Hover tooltips for skill tree implemented
