--- Game/Assets/Scenes/SkillTree/Scripts/HoverHandler.cs
@@ -1,60 +1,0 @@
-using UnityEngine;
-using UnityEngine.EventSystems;
-using UnityEngine.UI;
-using TMPro;
-
-public class HoverHandler : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler{
-
-
-    public GameObject hoverPanelPrefab;
-    GameObject hoverPanelInstance;
-    public Vector2 offset;
-
-    Player player;
-
-    void Init() {
-
-    }
-
-    void Awake() {
-
-        Init();
-
-    }
-
-    public void OnPointerEnter(PointerEventData eventData) {
-
-        if (hoverPanelPrefab != null && hoverPanelInstance == null) {
-
-            hoverPanelInstance = Instantiate(hoverPanelPrefab, transform);
-
-            string skillName = GetComponentInChildren<TMP_Text>().text;
-            hoverPanelInstance.GetComponent<HoverPanelContent>().SetTitle(skillName);
-
-            //skillDesc =
-            //hoverPanelInstance.transform.GetChild(2).gameObject.GetComponent<TMP_Text>.text = skillDesc;
-
-            RectTransform buttonRect = GetComponent<RectTransform>();
-            RectTransform panelRect = hoverPanelInstance.GetComponent<RectTransform>();
-
-            Vector3[] buttonCorners = new Vector3[4];
-            buttonRect.GetWorldCorners(buttonCorners);
-
-            Vector3 targetPos = buttonCorners[0] + new Vector3(offset.x, offset.y, 0);
-
-            panelRect.position = targetPos;
-        }
-
-    }
-
-    public void OnPointerExit(PointerEventData eventData) {
-
-        if (hoverPanelInstance != null) {
-
-            Destroy(hoverPanelInstance);
-
-        }
-
-    }
-
-}
