--- Game/Assets/Scenes/SkillTree/Scripts/HoverHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7b8bdf68844d842e29cde6c485540fc9
